requirements.txt
streamlit
langchain
PyMuPDF
faiss-cpu
requests
numpy
tiktoken
sentence-transformers